Return Result from main
main is allowed to return types that implement the
std::process::Termination trait, mainly Result.  If main returns Err(_),
the return code is set to indicate the failure and the Debug
representation of the error is printed to stderr.  anyhow::Result’s
Debug implementation is designed so that it can be used with this
feature.

This patch makes use of this feature by returning anyhow::Result from
main instead of handling the error manually.