Remove unnecessary string copies
In find_renames, we already have access to owned strings.  If we change
the arguments from references to vectors to owned vectors, we can just
move the existing strings into the created Rename structs instead of
copying the strings.

(Alternatively, we could drop the string copies in main and just pass
&[&str] to find_renames.)